配送状況をコンポーネント化
